fix: enable HMAC webhook signature verification
- Implement unified `verifyWebhookSignature` helper in `worker/lib/hmac.ts`
- Enable and enforce signature verification in `worker/routes/email.ts`
- Update `worker/routes/webhooks/incoming.ts` to return 401 for missing headers
- Add `WEBHOOK_SECRET` to environment configuration
- Add documentation for webhook signature generation in `docs/WEBHOOKS.md`
- Update unit tests to reflect 401 status for unauthorized webhooks

The fix addresses a critical security vulnerability where webhook signatures were either not checked or returning incorrect status codes on failure. Replay protection is included via timestamp validation with a 5-minute tolerance.
